to_do/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-appconf==1.0.5
django-compressor==4.4
django-libsass==0.9
django-sass==1.1.0
django-sass-processor==1.2.2
django-six==1.0.5
libsass==0.22.0
Pillow==10.0.0
python-dotenv==1.0.0
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
